refactor: seperated session functions from basic redis functionality
